requirements.txt
absl-py>=1.4.0
ansible>=8.2.0
ansible-core>=2.15.2
anyio>=3.7.0
argcomplete>=3.1.1
argon2-cffi>=21.3.0
argon2-cffi-bindings>=21.2.0
arrow>=1.2.3
asttokens>=2.2.1
attrs>=23.1.0
Automat>=22.10.0
autopep8>=2.0.2
Babel>=2.12.1
backcall>=0.2.0
bcrypt>=4.0.1
beautifulsoup4>=4.12.2
bleach>=6.0.0
blessed>=1.20.0
blivet>=3.8.0
bpyutils>=0.5.8
brightness-controller-linux>=2.4
cachetools>=5.3.0
canonicaljson>=2.0.0
certifi>=2023.5.7
cffi>=1.15.1
chardet>=5.1.0
charset-normalizer>=3.1.0
click>=8.1.6
click-plugins>=1.1.1
cligj>=0.7.2
cmake>=3.26.3
colorama>=0.4.6
comm>=0.1.3
conda-package-handling>=2.1.0
conda_package_streaming>=0.8.0
constantly>=15.1.0
contourpy>=1.0.7
cryptography>=41.0.2
cson>=0.8
cycler>=0.11.0
Cython>=0.29.35
cytoolz>=0.12.2
dasbus>=1.7
debugpy>=1.6.7
decorator>=5.1.1
defusedxml>=0.7.1
distlib>=0.3.7
distro>=1.8.0
dnspython>=2.4.0
email-validator>=2.0.0.post2
executing>=1.2.0
fastjsonschema>=2.17.1
file-magic>=0.4.1
filelock>=3.12.0
Fiona>=1.9.4.post1
flatbuffers>=23.5.9
fonttools>=4.39.4
fqdn>=1.5.1
frozendict>=2.3.8
geog>=0.0.2
google-auth>=2.19.0
google-auth-oauthlib>=1.0.0
googledrivedownloader>=0.4
grpcio>=1.54.2
h11>=0.14.0
HTMLParser>=0.0.2
httpcore>=0.17.3
idna>=3.4
ipykernel>=6.23.1
ipython>=8.13.2
ipython-genutils>=0.2.0
ipywidgets>=8.0.6
isoduration>=20.11.0
jedi>=0.18.2
Jinja2>=3.1.2
joblib>=1.2.0
jsonpointer>=2.3
jsonschema>=4.17.3
jupyter>=1.0.0
jupyter-console>=6.6.3
jupyter-events>=0.6.3
jupyter_client>=8.2.0
jupyter_core>=5.3.0
jupyter_server>=2.6.0
jupyter_server_terminals>=0.4.4
jupyterlab-pygments>=0.2.2
jupyterlab-widgets>=3.0.7
kiwisolver>=1.4.4
libclang>=16.0.0
libsumo>=1.17.0
lit>=16.0.5
littleutils>=0.2.2
lxml>=4.9.2
Markdown>=3.4.3
MarkupSafe>=2.1.2
matplotlib>=3.7.1
matplotlib-inline>=0.1.6
mistune>=2.0.5
monotonic>=1.6
mpmath>=1.3.0
nbclassic>=1.0.0
nbclient>=0.8.0
nbconvert>=7.5.0
nbformat>=5.9.0
nest-asyncio>=1.5.6
networkx>=3.1
notebook>=6.5.4
notebook_shim>=0.2.3
numpy>=1.24.3
nvidia-cublas-cu11>=11.10.3.66
nvidia-cuda-cupti-cu11>=11.7.101
nvidia-cuda-nvrtc-cu11>=11.7.99
nvidia-cuda-runtime-cu11>=11.7.99
nvidia-cudnn-cu11>=8.5.0.96
nvidia-cufft-cu11>=10.9.0.58
nvidia-curand-cu11>=10.2.10.91
nvidia-cusolver-cu11>=11.4.0.1
nvidia-cusparse-cu11>=11.7.4.91
nvidia-nccl-cu11>=2.14.3
nvidia-nvtx-cu11>=11.7.91
oauthlib>=3.2.2
ordereddict>=1.1
overrides>=7.3.1
packaging>=23.1
pandas>=2.0.1
pandocfilters>=1.5.0
parso>=0.8.3
pexpect>=4.8.0
pickleshare>=0.7.5
Pillow>=9.5.0
pip-review>=1.3.0
pipupgrade>=1.12.0
platformdirs>=3.5.1
ply>=3.11
prometheus-client>=0.17.0
prompt-toolkit>=3.0.38
protobuf>=4.23.1
psutil>=5.9.5
ptyprocess>=0.7.0
pure-eval>=0.2.2
pyasn1>=0.5.0
pyasn1-modules>=0.3.0
pycodestyle>=2.10.0
pycparser>=2.21
pyg-lib>=0.2.0+pt20cu118
Pygments>=2.15.1
pyparsing>=3.0.9
PyQt5>=5.15.9
PyQt5-Qt5>=5.15.2
PyQt5-sip>=12.12.1
pyrsistent>=0.19.3
python-dateutil>=2.8.2
python-gflags>=3.1.2
python-json-logger>=2.0.7
pytz>=2023.3
pyudev>=0.24.1
PyYAML>=6.0
pyzmq>=25.0.2
qtconsole>=5.4.3
QtPy>=2.3.1
requests>=2.31.0
requests-oauthlib>=1.3.1
resolvelib>=1.0.1
rfc3339-validator>=0.1.4
rfc3986-validator>=0.1.1
rsa>=4.9
scikit-learn>=1.2.2
scipy>=1.10.1
seaborn>=0.12.2
Send2Trash>=1.8.2
six>=1.16.0
sniffio>=1.3.0
soupsieve>=2.4.1
speg>=0.3
stack-data>=0.6.2
sumolib>=1.17.0
sympy>=1.12
tensorboard>=2.13.0
tensorboard-data-server>=0.7.0
tensorboard-plugin-wit>=1.8.1
terminado>=0.17.1
threadpoolctl>=3.1.0
tinycss2>=1.2.1
toolz>=0.12.0
torch>=2.0.1
torch-cluster>=1.6.1+pt20cu118
torch-geometric>=2.3.1
torch-geometric-temporal>=0.54.0
torch-scatter>=2.1.1+pt20cu118
torch-sparse>=0.6.17+pt20cu118
torch-spline-conv>=1.2.2+pt20cu118
torch-tb-profiler>=0.4.1
torchaudio>=2.0.2
torchvision>=0.15.2
tornado>=6.3.2
tqdm>=4.65.0
traci>=1.17.0
traitlets>=5.9.0
triton>=2.0.0
typing_extensions>=4.6.2
tzdata>=2023.3
uri-template>=1.2.0
urllib3>=1.26.16
wcwidth>=0.2.6
webcolors>=1.13
webencodings>=0.5.1
websocket-client>=1.5.3
Werkzeug>=2.3.4
wget>=3.2
widgetsnbextension>=4.0.7
zstandard>=0.21.0
